Clean up module exports: move State to model/simple, explicit model/index exports
