+++ packages/core/src/model/index.ts
@@ -1,4 +1,23 @@
-export * from "./primitives.js";
-export * from "./transform.js";
-export * from "./simple.js";
-export * from "./carousel.js";
+export { createModel, type State } from "./simple.js";
+export {
+  createTransformReduce,
+  settleTransform,
+  type BoundsConfig,
+  type TransformConfig,
+  type TransformSnapTarget,
+  type TransformAction,
+  type TransformPrivateState,
+} from "./transform.js";
+export {
+  createCarouselModel,
+  type CarouselConfig,
+  type CarouselPublicState,
+  type CarouselAction,
+  type CarouselPrivateState,
+} from "./carousel.js";
+export {
+  type Transform,
+  type TransformVelocity,
+  computeDtMs,
+  applyScalePivot,
+} from "./primitives.js";
